+++ src/main/kotlin/com/coder/toolbox/views/CoderSettingsPage.kt
@@ -125,12 +125,6 @@ class CoderSettingsPage(
         TextType.General
     )
 
-    private val workspaceFilterField = TextField(
-        context.i18n.ptrl("Workspace filter (leave blank to include shared workspaces)"),
-        settings.workspaceFilter,
-        TextType.General,
-    )
-
     private lateinit var visibilityUpdateJob: Job
     override val fields: StateFlow<List<UiField>> = MutableStateFlow(
         listOf(
@@ -140,7 +134,6 @@ class CoderSettingsPage(
                 listOf(
                     useAppNameField,
                     disableAutostartField,
-                    workspaceFilterField,
                     httpLoggingField,
                 )
             ),
@@ -221,7 +214,6 @@ class CoderSettingsPage(
                     updateSshLogDir(sshLogDirField.contentState.value)
                     updateNetworkInfoDir(networkInfoDirField.contentState.value)
                     updateSshConfigOptions(sshExtraArgs.contentState.value)
-                    updateWorkspaceFilter(workspaceFilterField.contentState.value)
                 }
             }
         )
@@ -294,10 +286,6 @@ class CoderSettingsPage(
 
         networkInfoDirField.contentState.update {
             settings.networkInfoDir
-        }
-
-        workspaceFilterField.contentState.update {
-            settings.workspaceFilter
         }
 
         visibilityUpdateJob = context.cs.launch(CoroutineName("Signature Verification Fallback Setting")) {